Isolate function requiring build artifacts
